requirements.txt
dash==2.17.0
dash-bootstrap-components==1.6.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn
pandas
requests
PyPDF2
python-docx
openpyxl
sentence-transformers
google-generativeai
qdrant-client
qdrant-client
dash
dash-bootstrap-components
plotly
fireworks-ai
pdf2image
pytesseract
pillow